tags -> key value pairs code added
Added code to parse and dump a csv of key:value pairs stored in tags.
This way, if you want to track something, you can type “key:value” into
the tags and it will dump it out to csv along with the full data. This
way you can say things like “steps:10434” and track it over the year.